Maybe created update operation

+++ dspace-api/src/main/java/org/dspace/handle/HandleClarinServiceImpl.java
@@ -13,17 +13,22 @@ import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
+import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.Logger;
-import org.dspace.content.DSpaceObject;
+import org.dspace.authorize.AuthorizeException;
+import org.dspace.authorize.service.AuthorizeService;
 import org.dspace.content.MetadataFieldServiceImpl;
 import org.dspace.content.service.ItemService;
 import org.dspace.core.Context;
 import org.dspace.core.LogHelper;
+import org.dspace.handle.dao.HandleClarinDAO;
 import org.dspace.handle.dao.HandleDAO;
+import org.dspace.handle.external.HandleRest;
 import org.dspace.handle.service.HandleClarinService;
 import org.dspace.handle.service.HandleService;
 import org.dspace.services.ConfigurationService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.util.ObjectUtils;
 
 import static org.dspace.handle.external.ExternalHandleConstants.MAGIC_BEAN;
 
@@ -34,8 +39,6 @@ import static org.dspace.handle.external.ExternalHandleConstants.MAGIC_BEAN;
  */
 public class HandleClarinServiceImpl implements HandleClarinService {
 
-    static final String PREFIX_DELIMITER = "/";
-
     /**
      * log4j logger
      */
@@ -45,6 +48,9 @@ public class HandleClarinServiceImpl implements HandleClarinService {
     protected HandleDAO handleDAO;
 
     @Autowired(required = true)
+    protected HandleClarinDAO handleClarinDAO;
+
+    @Autowired(required = true)
     protected HandleService handleService;
 
     @Autowired(required = true)
@@ -52,6 +58,11 @@ public class HandleClarinServiceImpl implements HandleClarinService {
 
     @Autowired(required = true)
     protected ConfigurationService configurationService;
+
+    @Autowired(required = true)
+    protected AuthorizeService authorizeService;
+
+    static final String PREFIX_DELIMITER = "/";
 
     /**
      * Public Constructor
@@ -65,27 +76,25 @@ public class HandleClarinServiceImpl implements HandleClarinService {
     }
 
     @Override
-    public List<Handle> findAllExternalHandles(Context context) throws SQLException {
-        // fetch all handles which contains `@magicLindat` string from the DB
-        return handleDAO.findAll(context, Handle.class)
-                .stream()
-                .filter(handle -> Objects.nonNull(handle))
-                .filter(handle -> Objects.nonNull(handle.getUrl()))
-                .filter(handle -> handle.getUrl().contains(MAGIC_BEAN))
-                .collect(Collectors.toList());
-    }
-
-    @Override
     public Handle findByID(Context context, int id) throws SQLException {
         return handleDAO.findByID(context, Handle.class, id);
     }
 
     @Override
-    public Handle createHandle(Context context, String handleStr, DSpaceObject dSpaceObject, String url)
-            throws SQLException {
-        String handleId = null;
+    public Handle findByHandle(Context context, String handle) throws SQLException {
+        return handleDAO.findByHandle(context, handle);
+    }
 
-        //Do we generate the new handleId generated or use entered handleStr by user?
+    @Override
+    public Handle createExternalHandle(Context context, String handleStr, String url)
+            throws SQLException, AuthorizeException {
+        // Check authorisation: Only admins may create DC types
+        if (!authorizeService.isAdmin(context)) {
+            throw new AuthorizeException(
+                    "Only administrators may modify the handle registry");
+        }
+        String handleId = null;
+        //Do we want to generate the new handleId or use entered handleStr by user?
         if (handleStr != null) {
             //we use handleStr entered by use
             handleId = handleStr;
@@ -96,27 +105,21 @@ public class HandleClarinServiceImpl implements HandleClarinService {
 
         Handle handle = handleDAO.create(context, new Handle());
 
-        //set handle depending on handleId
-        handle.setHandle(handleId);
-        //only if dspace object exists
-        if (dSpaceObject != null) {
-            handle.setDSpaceObject(dSpaceObject);
-            handle.setResourceTypeId(dSpaceObject.getType());
-        }
+        log.debug("Created new external Handle with handle " + handleId);
 
-        if (url != null) {
-            handle.setUrl(url);
-        }
-
-        handleDAO.save(context, handle);
-
-        log.debug("Created new Handle with handle " + handleId);
+        //set handle and url in created handle
+        setHandleAndUrlOfHandleObject(context, handle, handleStr, url);
 
         return handle;
     }
 
     @Override
-    public void delete(Context context, Handle handle) throws SQLException {
+    public void delete(Context context, Handle handle) throws SQLException, AuthorizeException {
+        // Check authorisation: Only admins may create DC types
+        if (!authorizeService.isAdmin(context)) {
+            throw new AuthorizeException(
+                    "Only administrators may modify the handle registry");
+        }
         //delete handle
         handleDAO.delete(context, handle);
 
@@ -125,7 +128,12 @@ public class HandleClarinServiceImpl implements HandleClarinService {
     }
 
     @Override
-    public void save(Context context, Handle handle) throws SQLException {
+    public void save(Context context, Handle handle) throws SQLException, AuthorizeException {
+        // Check authorisation: Only admins may create DC types
+        if (!authorizeService.isAdmin(context)) {
+            throw new AuthorizeException(
+                    "Only administrators may modify the handle registry");
+        }
         //save handle
         handleDAO.save(context, handle);
 
@@ -137,30 +145,34 @@ public class HandleClarinServiceImpl implements HandleClarinService {
 
     @Override
     public void update(Context context, Handle handleObject, String newHandle,
-                       DSpaceObject dso, String newUrl)
-            throws SQLException {
-        //set all handle attributes
-        handleObject.setHandle(newHandle);
-        handleObject.setDSpaceObject(dso);
-        //if dspace object is null, set resource type id to null
-        if (dso != null) {
-            //resource type id is type od dspace object
-            handleObject.setResourceTypeId(dso.getType());
-        } else {
-            handleObject.setResourceTypeId(null);
+                       String newUrl)
+            throws SQLException, AuthorizeException {
+        // Check authorisation: Only admins may create DC types
+        if (!authorizeService.isAdmin(context)) {
+            throw new AuthorizeException(
+                    "Only administrators may modify the handle registry");
         }
-        if (newUrl != null) {
-            handleObject.setUrl(newUrl);
-        }
-
-        this.save(context, handleObject);
+        //set handle and url in handle
+        setHandleAndUrlOfHandleObject(context, handleObject, newHandle, newUrl);
 
         log.info(LogHelper.getHeader(context, "update_handle",
                 "handle_id=" + handleObject.getID()));
     }
 
     @Override
-    public void setPrefix(Context context, String newPrefix, String oldPrefix) throws SQLException {
+  //  @PreAuthorize("hasAuthority('ADMIN')")
+    public void setPrefix(Context context, String newPrefix, String oldPrefix) throws SQLException,
+            AuthorizeException {
+        // Check authorisation: Only admins may create DC types
+        if (!authorizeService.isAdmin(context)) {
+            throw new AuthorizeException(
+                    "Only administrators may modify the handle registry");
+        }
+        //control, if are new and old prefix entered
+        if (ObjectUtils.isEmpty(newPrefix) && StringUtils.isBlank(newPrefix) &&
+                ObjectUtils.isEmpty(oldPrefix) && StringUtils.isBlank(oldPrefix)) {
+            throw new NullPointerException("Cannot set prefix. Required fields are empty.");
+        }
         //get handle prefix
         String prefix = handleService.getPrefix();
         //set prefix only if not equal to old prefix
@@ -170,6 +182,8 @@ public class HandleClarinServiceImpl implements HandleClarinService {
                 //prefix has not changed
                 throw new RuntimeException("error while trying to set handle prefix");
             }
+        } else {
+            throw new RuntimeException("Cannot set prefix. Entered prefix does not match with ");
         }
 
         log.info(LogHelper.getHeader(context, "set_handle_prefix",
@@ -203,7 +217,7 @@ public class HandleClarinServiceImpl implements HandleClarinService {
             //internal handle
             //create url for internal handle
             url = configurationService.getProperty("dspace.ui.url")
-                    + "/handle/" + handle;
+                    + "/handle/" + handleStr;
         } else {
             //external handle
             url = handle.getUrl();
@@ -212,23 +226,6 @@ public class HandleClarinServiceImpl implements HandleClarinService {
         log.debug("Resolved {} to {}", handle, url);
 
         return url;
-    }
-
-    @Override
-    public List<org.dspace.handle.external.Handle> convertHandleWithMagicToExternalHandle(List<Handle> magicHandles) {
-        List<org.dspace.handle.external.Handle> externalHandles = new ArrayList<>();
-        for (org.dspace.handle.Handle handleWithMagic: magicHandles) {
-            externalHandles.add(new org.dspace.handle.external.Handle(handleWithMagic.getHandle(), handleWithMagic.getUrl()));
-        }
-
-        return externalHandles;
-    }
-
-    /**
-     * Returns complete handle made from prefix and suffix
-     */
-    public String completeHandle(String prefix, String suffix) {
-        return prefix + PREFIX_DELIMITER + suffix;
     }
 
     /**
@@ -247,5 +244,94 @@ public class HandleClarinServiceImpl implements HandleClarinService {
         return handlePrefix + (handlePrefix.endsWith("/") ? "" : "/") + handleSuffix.toString();
     }
 
+    /**
+     * Set handle and url of handle object.
+     * It is not possible to change internal handle to external handle or
+     * external handle to internal handle.
+     *
+     * @param context       DSpace context object
+     * @param handleObject       handle object
+     * @param newHandle     new string handle
+     * @param newUrl     new url
+     * @throws SQLException if database error
+     * @throws AuthorizeException if authorization error
+     */
+    private void setHandleAndUrlOfHandleObject(Context context, Handle handleObject, String newHandle,
+                                           String newUrl) throws SQLException, AuthorizeException {
+        //set handle
+        handleObject.setHandle(newHandle);
+        //if it is internal handle, do nothing with url
+        if (newUrl != null) {
+            //set url only if is not empty
+            //when you add null to String, it converts null to "null"
+            if (!(ObjectUtils.isEmpty(newUrl)) && !(StringUtils.isBlank(newUrl)) &&
+                newUrl != "null") {
+                handleObject.setUrl(newUrl);
+            } else {
+                throw new RuntimeException("Cannot change handle and url of handle object.");
+            }
+        }
 
+        this.save(context, handleObject);
+
+        log.info(LogHelper.getHeader(context, "Set handle and url of handle object.",
+                "handle_id=" + handleObject.getID()));
+    }
+
+    @Override
+    public List<org.dspace.handle.external.Handle> convertHandleWithMagicToExternalHandle(List<Handle> magicHandles) {
+        List<org.dspace.handle.external.Handle> externalHandles = new ArrayList<>();
+        for (org.dspace.handle.Handle handleWithMagic: magicHandles) {
+            externalHandles.add(new org.dspace.handle.external.Handle(handleWithMagic.getHandle(), handleWithMagic.getUrl()));
+        }
+
+        return externalHandles;
+    }
+
+    @Override
+    public List<HandleRest> convertExternalHandleToHandleRest(List<org.dspace.handle.external.Handle> externalHandles) {
+        List<HandleRest> externalHandleRestList = new ArrayList<>();
+        for (org.dspace.handle.external.Handle externalHandle: externalHandles) {
+            HandleRest externalHandleRest = new HandleRest();
+
+            externalHandleRest.setHandle(externalHandle.getHandle());
+            externalHandleRest.setUrl(externalHandle.url);
+            externalHandleRest.setTitle(externalHandle.title);
+            externalHandleRest.setSubprefix(externalHandle.subprefix);
+            externalHandleRest.setReportemail(externalHandle.reportemail);
+            externalHandleRest.setRepository(externalHandle.repository);
+            externalHandleRest.setSubmitdate(externalHandle.submitdate);
+
+            externalHandleRestList.add(externalHandleRest);
+        }
+
+        return externalHandleRestList;
+    }
+
+    /**
+     * Returns complete handle made from prefix and suffix
+     */
+    public String completeHandle(String prefix, String suffix) {
+        return prefix + PREFIX_DELIMITER + suffix;
+    }
+
+    @Override
+    public List<Handle> findAllExternalHandles(Context context) throws SQLException {
+        // fetch all handles which contains `@magicLindat` string from the DB
+        return handleDAO.findAll(context, Handle.class)
+                .stream()
+                .filter(handle -> Objects.nonNull(handle))
+                .filter(handle -> Objects.nonNull(handle.getUrl()))
+                .filter(handle -> handle.getUrl().contains(MAGIC_BEAN))
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public Handle findHandleByHandle(Context context, String handle) throws SQLException {
+        if (handle == null) {
+            throw new IllegalArgumentException("Handle is null");
+        }
+
+        return handleDAO.findByHandle(context, handle);
+    }
 }
